refactor tests dist
