+++ tests/test_dist/test_diloco.py
@@ -1,0 +1,1 @@
+"""test Diloco. Need 4 gpus to run this tests"""
